🧹 remove unnecessary class
