+++ lib/ConsumptionDisplay/stories.js
@@ -1,5 +1,5 @@
 import dayjs from 'dayjs'
-import ConsumptionDisplay from './ConsumptionDisplay'
+import ConsumptionDisplay from '.'
 
 export default {
   title: 'Energy components/ConsumptionDisplay',
