+++ lib/SummaryDisplay/index.jsx
