+++ lib/SummaryDisplay/stories.js
@@ -1,4 +1,4 @@
-import SummaryDisplay from "./SummaryDisplay";
+import SummaryDisplay from ".";
 
 export default {
   title: 'Base Components/SummaryDisplay',
